requirements-tests.txt
coverage
mock
nose
requests-mock
tox
